requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
dj-static==0.0.6
Django==4.0.6
django-stdimage==6.0.1
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
requests==2.28.1
sqlparse==0.4.2
static3==0.7.0
stripe==3.5.0
urllib3==1.26.12
